Port to v8 and to new API
