+++ account_journal_always_check_date/i18n/account_journal_always_check_date.pot
